Split parser into several layered sub-languages
Each sub-language should be easier to work with, understand, and reason about.
I'm not sure if there's a lot we can do to make use of the sub-languages, but
it still seems worthwhile to split them apart. Note that I have not (yet)
separated the AST classes in any way; they can only be accessed through the
corresponding grammars, so it didn't seem a worthwhile effort.

The only other programming language I'm aware of that has layered sub-languages
is [Noether](https://www.infoq.com/presentations/noether/). The layering in
Stone was partly inspired by Noether. Noether has also been an inspiration with
respect to the philosophy of not allowing operations to succeed with an
incorrect answer.