Ruby implementation of Church Booleans
I wrote this in Ruby, to get a better understanding of how all the functions
could be written simply, so that I can port it to Stone. I used stabby lambdas,
as they quite closely align with what I've already got working in Stone.
I also had some fun using Unicode symbols as the names of the functions.

Thanks to Amos for recommending Church Booleans to me.