Implement Booleans using Church encoding
I ported this over from my Ruby implementation. (See the previous commit.)
It's a little premature to test this, but I want to get as much library
code written in Stone as I can, to help drive the language design and
implementation. I believe all that's left to test this is making a Class
definition work. (There's a type definition too, but I can comment that
out for now.)

The really nice thing about this implementation is that it requires
virtually NO dependencies. It's mostly just function definitions and
function calls. Unfortunately, that won't really work long-term, as
we can't embed an object with 0 bytes into another data structure.
(Maybe just having a pointer to the TRUE and FALSE functionswould work?)
We'll likely change this to use an opaque Enum to solve that problem.
But as far as having something written in pure Stone that gives us
almost everything we could want from Booleans, I'm really pleased.

Thanks to @adkron for suggesting Church Booleans.