Implement closures
I'm confident that I got it right this time. (I had previously tried, but
failed to get some things right.) Note that as everything in Stone is read-only
(at least for now), variables in outer scopes that are referenced in inner
scopes cannot be mutated. Therefore this is a much simpler implementation than
if we had to be able to keep the state of a variable once its scope disappears.
All we have to do is determine which variables in a function are free (no bound
to anything *within* the function) and capture their values from the enclosing
scope at the time the function is created.

Also did a little bit of cleanup to make RuboCop happier.