BREAKING CHANGE: Use `λ` or `->` to create a function
The syntax is now `λ(x) {}`. The `(x) => {}` syntax is no longer valid.
This is to make Stone's syntax look more like psuedo-code, for those familiar
with Lambda Calculus.

There's also a "stabby lambda" alternative, if `λ` is too difficult to type.
This syntax matches Ruby syntax (although it will create a "proc" in Ruby,
not a function).