Change property accessed to allow inheritance
This is a pretty extensive change, that allows us to define a property in a
superclass that will be available in subclasses, unless overridden.

Note that most of the work is done in the Value class, as all Values can have
properties. (And they'll all have a few, such as `type` and `null?`.)

Value is quite similar to Ruby's BasicObject.