Remove experiments from Verfification::Suite
I believe that there's a better way to write this, and a blog article to write
once I figure out what that is. I experimented with several different ways, but
need to spend some time to figure out which way to go.

In the meantime, I'm removing the experiments of different ways to try to code
this. I can come back to the `clean-up-Verification-Suite` branch later.