Identifiers should not allow lambda character (`λ`)
We're reserving that for functions.

We might be able to make `λ` a built-in in the future, but I'm not convinced
that that is likely to work out. It would nice to be able to make that work,
though. Then we can allow `λ` in identifiers, and define it in the prelude:

~~~ stone
λ := __BUILTIN__.lambda
~~~