Allow comments (from `#` to EOL)
We include comments in the parse tree, because the parser will be used for code
formatting and other purposes. Comments do not effect the AST.

NOTE: Comment nodes in the parse tree include the trailing EOL in their
captured text.

The grammar is slightly complicated by the fact that a line with a comment may
or may not include a statement before the comment.